add authentication fie and util file
