+++ .gitignore
@@ -1,2 +1,4 @@
 /Chun_env
 /.env
+/__pycache__
+/outline 1.txt
