+++ Data/History.csv
@@ -19,3 +19,9 @@ Male,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Ye
 
 Male,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Yes,Electronic check
 Male,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Yes,Electronic check
+Male,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Yes,Electronic check
+Male,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Yes,Electronic check
+Female,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Yes,Electronic check
+Male,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Yes,Electronic check
+Male,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Yes,Electronic check
+Male,Yes,Yes,1,0.0,0.0,0.0,Yes,Yes,DSL,Yes,Yes,Yes,Yes,Yes,Yes,Month-to-Month,Yes,Electronic check
